chore: Generate secondary indexes during migration
