+++ dozer-api/src/cache_builder/mod.rs
@@ -2,20 +2,17 @@ use std::{path::Path, sync::Arc};
 
 use crate::grpc::types_helper;
 use dozer_cache::dozer_log::reader::LogReader;
-use dozer_cache::errors::IndexError;
 use dozer_cache::{
     cache::{CacheRecord, CacheWriteOptions, RwCache, RwCacheManager, UpsertResult},
     errors::CacheError,
 };
 use dozer_types::epoch::ExecutorOperation;
 use dozer_types::indicatif::MultiProgress;
-use dozer_types::models::api_endpoint::{
-    FullText, SecondaryIndex, SecondaryIndexConfig, SortedInverted,
-};
+use dozer_types::types::IndexDefinition;
 use dozer_types::{
     grpc_types::types::Operation as GrpcOperation,
     log::error,
-    types::{Field, FieldDefinition, FieldType, IndexDefinition, Operation, Record, Schema},
+    types::{Field, Operation, Record, Schema},
 };
 use futures_util::{
     future::{select, Either},
@@ -27,7 +24,7 @@ use tokio::{runtime::Runtime, sync::broadcast::Sender};
 pub async fn create_cache(
     cache_manager: &dyn RwCacheManager,
     schema: Schema,
-    secondary_index_config: &SecondaryIndexConfig,
+    secondary_indexes: Vec<IndexDefinition>,
     runtime: Arc<Runtime>,
     cancel: impl Future<Output = ()> + Unpin,
     log_path: &Path,
@@ -35,8 +32,6 @@ pub async fn create_cache(
     operations_sender: Option<(String, Sender<GrpcOperation>)>,
     multi_pb: Option<MultiProgress>,
 ) -> Result<(String, impl FnOnce() -> Result<(), CacheError>), CacheError> {
-    // Create secondary indexes
-    let secondary_indexes = generate_secondary_indexes(&schema.fields, secondary_index_config)?;
     // Create the cache.
     let cache = cache_manager.create_cache(schema.clone(), secondary_indexes, write_options)?;
     let name = cache.name().to_string();
@@ -128,74 +123,6 @@ fn build_cache(
     }
 
     Ok(())
-}
-
-fn generate_secondary_indexes(
-    field_definitions: &[FieldDefinition],
-    config: &SecondaryIndexConfig,
-) -> Result<Vec<IndexDefinition>, CacheError> {
-    let mut result = vec![];
-
-    // Create default indexes unless skipped.
-    for (index, field) in field_definitions.iter().enumerate() {
-        if config.skip_default.contains(&field.name) {
-            continue;
-        }
-
-        match field.typ {
-            // Create sorted inverted indexes for these fields
-            FieldType::UInt
-            | FieldType::U128
-            | FieldType::Int
-            | FieldType::I128
-            | FieldType::Float
-            | FieldType::Boolean
-            | FieldType::Decimal
-            | FieldType::Timestamp
-            | FieldType::Date
-            | FieldType::Point
-            | FieldType::Duration => result.push(IndexDefinition::SortedInverted(vec![index])),
-
-            // Create sorted inverted and full text indexes for string fields.
-            FieldType::String => {
-                result.push(IndexDefinition::SortedInverted(vec![index]));
-                result.push(IndexDefinition::FullText(index));
-            }
-
-            // Skip creating indexes
-            FieldType::Text | FieldType::Binary | FieldType::Json => (),
-        }
-    }
-
-    // Create requested indexes.
-    fn field_index_from_field_name(
-        fields: &[FieldDefinition],
-        field_name: &str,
-    ) -> Result<usize, IndexError> {
-        fields
-            .iter()
-            .position(|field| field.name == field_name)
-            .ok_or(IndexError::UnknownFieldName(field_name.to_string()))
-    }
-    for create in &config.create {
-        if let Some(index) = &create.index {
-            match index {
-                SecondaryIndex::SortedInverted(SortedInverted { fields }) => {
-                    let fields = fields
-                        .iter()
-                        .map(|field| field_index_from_field_name(field_definitions, field))
-                        .collect::<Result<Vec<_>, _>>()?;
-                    result.push(IndexDefinition::SortedInverted(fields));
-                }
-                SecondaryIndex::FullText(FullText { field }) => {
-                    let field = field_index_from_field_name(field_definitions, field)?;
-                    result.push(IndexDefinition::FullText(field));
-                }
-            }
-        }
-    }
-
-    Ok(result)
 }
 
 async fn next_op_with_cancel<F: Future<Output = ()> + Unpin>(
